install.sh: Be sensitive to which Go version is in use.
Older versions of Go will not "go install" with a version label, but require
that you use "go get" for that purpose (e.g., installing binaries).

Newer versions of Go complain if you do this, and print a deprecation warning
that you should be using "go install".

To avoid the failure and the deprecation warning, check what version we are
using and do the needful.